requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiohttp==3.9.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0 